Moved old code to old_pipeline, so it is clear that it is not needed.
