+++ .gitignore
@@ -1,4 +1,4 @@
-secrets.py
+old_pipeline/secrets.py
 
 __pycache__/
