+++ old_pipeline/README.md
@@ -1,0 +1,7 @@
+
+# Pipeline Directory
+
+Most the files in this directory are OBE (overcome by events).  They are using 
+a home-grown scaling solution on AWS, which consisted of maintaining a list of 
+scenes, utilities to manage remote machines and processes, and performer-specific
+code to run the evaluation.
