+++ old_pipeline/mcs_test_runner.py
@@ -4,7 +4,7 @@
 # running on a remote machine. Also check for timeout
 #
 
-from pipeline import util
+from old_pipeline import util
 
 
 class McsTestRunner:
